Potential fix for code scanning alert no. 43: Multiplication result converted to larger type
To fix the problem, we need to ensure that the multiplication is performed using the larger integer type (uint64_t) to avoid overflow. This can be done by casting one of the operands to uint64_t before performing the multiplication. This way, the multiplication will be done in the larger type, preventing overflow.

 - Cast one of the operands to uint64_t before the multiplication.
 - Specifically, cast enc->mb_w_ to uint64_t to ensure the multiplication is done using uint64_t.
 - The change will be made on line 782 in the file libwebp-1.0.3/src/enc/frame_enc.c.

+++ libwebp-1.0.3/src/enc/frame_enc.c
@@ -779,7 +779,7 @@ int VP8EncTokenLoop(VP8Encoder* const enc) {
   VP8EncIterator it;
   VP8EncProba* const proba = &enc->proba_;
   const VP8RDLevel rd_opt = enc->rd_opt_level_;
-  const uint64_t pixel_count = enc->mb_w_ * enc->mb_h_ * 384;
+  const uint64_t pixel_count = (uint64_t)enc->mb_w_ * enc->mb_h_ * 384;
   PassStats stats;
   int ok;
